parei num erro de classes

+++ src/vector2.cpp
@@ -1,6 +1,6 @@
 #include <vector2.h>
 
-Vector2::Vector2(int x_p, int y_p) {
+Vector2::Vector2(int64_t x_p, int64_t y_p) {
     x = x_p;
     y = y_p;
 }
